# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface WaitlistRegistrationInput {
  /**
     * @maxLength 320
     * @pattern ^[^\s@]+@[^\s@]+\.[^\s@]+$
     */
  email: string;
  privacyAccepted: true;
  /** @maxLength 80 */
  source?: string;
}

export interface WaitlistRegistrationResponse {
  accepted: boolean;
  message: string;
  alreadyConfirmed?: boolean;
  /** True only when the email has confirmed a place in the first 500. */
  eligible?: boolean;
  /**
     * @minimum 0
     * @maximum 100
     */
  discountPercent?: number;
}

export type WaitlistEligibilityPlanPrices = {
  monthly: number;
  annual: number;
};

export interface WaitlistEligibility {
  confirmed: boolean;
  eligible: boolean;
  /** @nullable */
  confirmationRank?: number | null;
  /**
     * @minimum 0
     * @maximum 100
     */
  discountPercent: number;
  planPrices?: WaitlistEligibilityPlanPrices;
}

export interface SecuritySchemes {}