# lib/api-zod/src/generated/types/waitlistEligibility.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WaitlistEligibilityPlanPrices } from './waitlistEligibilityPlanPrices';

export interface WaitlistEligibility {
  confirmed: boolean;
  eligible: boolean;
  /** @nullable */
  confirmationRank?: number | null;
  /**
     * @minimum 0
     * @maximum 100
     */
  discountPercent: number;
  planPrices?: WaitlistEligibilityPlanPrices;
}