# lib/api-zod/src/generated/types/waitlistRegistrationResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WaitlistRegistrationResponse {
  accepted: boolean;
  message: string;
  alreadyConfirmed?: boolean;
  /** True only when the email has confirmed a place in the first 500. */
  eligible?: boolean;
  /**
     * @minimum 0
     * @maximum 100
     */
  discountPercent?: number;
}